requirements.txt

google-generativeai==0.1.0
pytest==7.4.0
deepeval==0.2.0
python-dotenv==1.0.0
allure-pytest==2.13.0
jsonschema==4.19.0
pandas==2.2.0
sqlalchemy==2.0.0
duckdb==0.8.1
